+++ packages/playground/src/components/App/App.js
@@ -21,11 +21,13 @@ var galleryReadyEvent = new Event('galleryReady');
 
 export function App() {
 
-  const {setDimentions, styleParams, setItems, items} = useGalleryContext();
+  const {setDimentions, styleParams, setItems, items, isSSR} = useGalleryContext();
   const [showSide, setShowSide] = useState(true);
-  const [fullscreenIdx, setFullscreenIdx] = useState(-1);
+  // const [fullscreenIdx, setFullscreenIdx] = useState(-1);
 
   setStyleParamsInUrl(styleParams);
+
+  // !isSSR && setIsSSR(true);
 
   const switchState = () => {
     const width = showSide ? window.innerWidth : window.innerWidth - SIDEBAR_WIDTH;
@@ -58,7 +60,7 @@ export function App() {
         setItems(getItems().concat(mixAndSlice(testItems, ITEMS_BATCH_SIZE)));
         break;
       case GALLERY_EVENTS.ITEM_ACTION_TRIGGERED: 
-        setFullscreenIdx(eventData.idx);
+        // setFullscreenIdx(eventData.idx);
         break;
       default: 
         // console.log({eventName, eventData});
@@ -68,7 +70,7 @@ export function App() {
 
   const container = {
     height: window.innerHeight,
-    width: window.innerWidth - (showSide ? SIDEBAR_WIDTH : 0)
+    width: isSSR ? '' : window.innerWidth - (showSide ? SIDEBAR_WIDTH : 0)
   }
 
   const getItems = () => {
@@ -95,8 +97,8 @@ export function App() {
           resizeMediaUrl={resizeMediaUrl}
         />
       </section>
-      <section className={['pro-fullscreen-wrapper', s.fullscreen].join(' ')} style={{...container, opacity: (fullscreenIdx >= 0 ? 1 : 0), pointerEvents: (fullscreenIdx >= 0 ? 'initial' : 'none')}}>
-      {/* fullscreenIdx >= 0 && (
+      {/* <section className={['pro-fullscreen-wrapper', s.fullscreen].join(' ')} style={{...container, opacity: (fullscreenIdx >= 0 ? 1 : 0), pointerEvents: (fullscreenIdx >= 0 ? 'initial' : 'none')}}>
+      fullscreenIdx >= 0 && (
           <ProFullscreen
             items={images}
             initialIdx={fullscreenIdx}
@@ -110,8 +112,8 @@ export function App() {
             deviceType={'desktop'}
             styles={styleParams}
           />
-        ) */}
-        </section>
+        )
+        </section> */}
     </main>
   );
 }