+++ packages/playground/src/hooks/useGalleryContext.js
@@ -36,6 +36,10 @@ export function useGalleryContext() {
     setContext({galleryReady});
   };
 
+  const setIsSSR = isSSR => {
+    setContext({isSSR});
+  }
+
   const res = {
     showSide: context.showSide,
     setShowSide,
@@ -47,6 +51,8 @@ export function useGalleryContext() {
     setItems,
     galleryReady: context.galleryReady,
     setGalleryReady,
+    isSSR: context.isSSR,
+    setIsSSR,
     dimensions: {
       width: context.galleryWidth,
       height: context.galleryHeight,